remove individual BucketProviders
