+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/dataformats/DatasetArrayBucketProvider.scala
@@ -2,24 +2,32 @@ package com.scalableminds.webknossos.datastore.dataformats
 
 import com.scalableminds.util.cache.AlfuCache
 import com.scalableminds.util.geometry.Vec3Int
-import com.scalableminds.util.tools.Fox
+import com.scalableminds.util.tools.{Fox, FoxImplicits}
 import com.scalableminds.webknossos.datastore.datareaders.DatasetArray
-import com.scalableminds.webknossos.datastore.models.datasource.{DataLayer, DataSourceId, ElementClass}
+import com.scalableminds.webknossos.datastore.datareaders.n5.N5Array
+import com.scalableminds.webknossos.datastore.datareaders.precomputed.PrecomputedArray
+import com.scalableminds.webknossos.datastore.datareaders.wkw.WKWArray
+import com.scalableminds.webknossos.datastore.datareaders.zarr.ZarrArray
+import com.scalableminds.webknossos.datastore.datareaders.zarr3.Zarr3Array
+import com.scalableminds.webknossos.datastore.datavault.VaultPath
+import com.scalableminds.webknossos.datastore.models.datasource.{DataFormat, DataLayer, DataSourceId, ElementClass}
 import com.scalableminds.webknossos.datastore.models.requests.DataReadInstruction
 import com.scalableminds.webknossos.datastore.storage.RemoteSourceDescriptorService
 import com.typesafe.scalalogging.LazyLogging
+import net.liftweb.common.Empty
 import ucar.ma2.{Array => MultiArray}
 
 import scala.concurrent.ExecutionContext
 
-abstract class DatasetArrayBucketProvider(dataLayer: DataLayer,
-                                          dataSourceId: DataSourceId,
-                                          val remoteSourceDescriptorServiceOpt: Option[RemoteSourceDescriptorService],
-                                          sharedChunkContentsCache: Option[AlfuCache[String, MultiArray]])
+class DatasetArrayBucketProvider(dataLayer: DataLayer,
+                                 dataSourceId: DataSourceId,
+                                 val remoteSourceDescriptorServiceOpt: Option[RemoteSourceDescriptorService],
+                                 sharedChunkContentsCache: Option[AlfuCache[String, MultiArray]])
     extends AbstractBucketProvider
+    with FoxImplicits
     with LazyLogging {
 
-  // by mag
+  // Cache the DatasetArrays of all mags of this layer
   private lazy val arrayHandleCache = AlfuCache[Vec3Int, DatasetArray](maxCapacity = 50)
 
   def load(readInstruction: DataReadInstruction)(implicit ec: ExecutionContext): Fox[Array[Byte]] =
@@ -54,6 +62,63 @@ abstract class DatasetArrayBucketProvider(dataLayer: DataLayer,
     } yield result
   }
 
-  def openDatasetArrayHandle(readInstruction: DataReadInstruction)(implicit ec: ExecutionContext): Fox[DatasetArray]
+  private def openDatasetArrayHandle(readInstruction: DataReadInstruction)(
+      implicit ec: ExecutionContext): Fox[DatasetArray] = {
+    val magLocatorOpt: Option[MagLocator] =
+      dataLayer.mags.find(_.mag == readInstruction.bucket.mag)
+
+    magLocatorOpt match {
+      case None => Fox.empty
+      case Some(magLocator) =>
+        remoteSourceDescriptorServiceOpt match {
+          case Some(remoteSourceDescriptorService: RemoteSourceDescriptorService) =>
+            for {
+              magPath: VaultPath <- remoteSourceDescriptorService.vaultPathFor(readInstruction.baseDir,
+                                                                               readInstruction.dataSource.id,
+                                                                               readInstruction.dataLayer.name,
+                                                                               magLocator)
+              chunkContentsCache <- sharedChunkContentsCache.toFox
+              datasetArray <- dataLayer.dataFormat match {
+                case DataFormat.zarr =>
+                  ZarrArray.open(magPath,
+                                 dataSourceId,
+                                 dataLayer.name,
+                                 magLocator.axisOrder,
+                                 magLocator.channelIndex,
+                                 dataLayer.additionalAxes,
+                                 chunkContentsCache)
+                case DataFormat.wkw =>
+                  WKWArray.open(magPath, dataSourceId, dataLayer.name, chunkContentsCache)
+                case DataFormat.n5 =>
+                  N5Array.open(magPath,
+                               dataSourceId,
+                               dataLayer.name,
+                               magLocator.axisOrder,
+                               magLocator.channelIndex,
+                               dataLayer.additionalAxes,
+                               chunkContentsCache)
+                case DataFormat.zarr3 =>
+                  Zarr3Array.open(magPath,
+                                  dataSourceId,
+                                  dataLayer.name,
+                                  magLocator.axisOrder,
+                                  magLocator.channelIndex,
+                                  dataLayer.additionalAxes,
+                                  chunkContentsCache)
+                case DataFormat.neuroglancerPrecomputed =>
+                  PrecomputedArray.open(magPath,
+                                        dataSourceId,
+                                        dataLayer.name,
+                                        magLocator.axisOrder,
+                                        magLocator.channelIndex,
+                                        dataLayer.additionalAxes,
+                                        chunkContentsCache)
+                case _ => Fox.failure(s"Cannot open ${dataLayer.dataFormat} layer “${dataLayer.name}” as DatasetArray")
+              }
+            } yield datasetArray
+          case None => Empty
+        }
+    }
+  }
 
 }
