--- webknossos-datastore/app/com/scalableminds/webknossos/datastore/dataformats/zarr3/Zarr3DataLayers.scala
@@ -1,72 +1,0 @@
-package com.scalableminds.webknossos.datastore.dataformats.zarr3
-
-import com.scalableminds.util.cache.AlfuCache
-import com.scalableminds.util.geometry.{BoundingBox, Vec3Int}
-import com.scalableminds.webknossos.datastore.dataformats.MagLocator
-import com.scalableminds.webknossos.datastore.models.datasource.{
-  AdditionalAxis,
-  Category,
-  CoordinateTransformation,
-  DataFormat,
-  DataLayerWithMagLocators,
-  DataSourceId,
-  ElementClass,
-  SegmentationLayer
-}
-import com.scalableminds.webknossos.datastore.models.datasource.LayerViewConfiguration.LayerViewConfiguration
-import com.scalableminds.webknossos.datastore.storage.RemoteSourceDescriptorService
-import play.api.libs.json.{Json, OFormat}
-import ucar.ma2.{Array => MultiArray}
-
-trait Zarr3Layer extends DataLayerWithMagLocators {
-
-  val dataFormat: DataFormat.Value = DataFormat.zarr3
-
-  def bucketProvider(remoteSourceDescriptorServiceOpt: Option[RemoteSourceDescriptorService],
-                     dataSourceId: DataSourceId,
-                     sharedChunkContentsCache: Option[AlfuCache[String, MultiArray]]) =
-    new Zarr3BucketProvider(this, dataSourceId, remoteSourceDescriptorServiceOpt, sharedChunkContentsCache)
-
-  def resolutions: List[Vec3Int] = mags.map(_.mag)
-
-  def lengthOfUnderlyingCubes(resolution: Vec3Int): Int = Int.MaxValue // Prevents the wkw-shard-specific handle caching
-
-  def numChannels: Option[Int] = Some(if (elementClass == ElementClass.uint24) 3 else 1)
-
-}
-
-case class Zarr3DataLayer(
-                           name: String,
-                           category: Category.Value,
-                           boundingBox: BoundingBox,
-                           elementClass: ElementClass.Value,
-                           mags: List[MagLocator],
-                           defaultViewConfiguration: Option[LayerViewConfiguration] = None,
-                           adminViewConfiguration: Option[LayerViewConfiguration] = None,
-                           coordinateTransformations: Option[List[CoordinateTransformation]] = None,
-                           override val numChannels: Option[Int] = Some(1),
-                           additionalAxes: Option[Seq[AdditionalAxis]] = None
-) extends Zarr3Layer
-
-object Zarr3DataLayer {
-  implicit val jsonFormat: OFormat[Zarr3DataLayer] = Json.format[Zarr3DataLayer]
-}
-
-case class Zarr3SegmentationLayer(
-                                   name: String,
-                                   boundingBox: BoundingBox,
-                                   elementClass: ElementClass.Value,
-                                   mags: List[MagLocator],
-                                   largestSegmentId: Option[Long] = None,
-                                   mappings: Option[Set[String]] = None,
-                                   defaultViewConfiguration: Option[LayerViewConfiguration] = None,
-                                   adminViewConfiguration: Option[LayerViewConfiguration] = None,
-                                   coordinateTransformations: Option[List[CoordinateTransformation]] = None,
-                                   override val numChannels: Option[Int] = Some(1),
-                                   additionalAxes: Option[Seq[AdditionalAxis]] = None
-) extends SegmentationLayer
-    with Zarr3Layer
-
-object Zarr3SegmentationLayer {
-  implicit val jsonFormat: OFormat[Zarr3SegmentationLayer] = Json.format[Zarr3SegmentationLayer]
-}
