+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/services/DataSourceService.scala
@@ -9,11 +9,8 @@ import com.scalableminds.util.time.Instant
 import com.scalableminds.util.tools.{Fox, FoxImplicits, JsonHelper}
 import com.scalableminds.webknossos.datastore.DataStoreConfig
 import com.scalableminds.webknossos.datastore.dataformats.MappingProvider
-import com.scalableminds.webknossos.datastore.dataformats.n5.N5Layer
-import com.scalableminds.webknossos.datastore.dataformats.precomputed.PrecomputedLayer
+import com.scalableminds.webknossos.datastore.dataformats.layers.{N5Layer, PrecomputedLayer, Zarr3Layer, ZarrLayer}
 import com.scalableminds.webknossos.datastore.dataformats.wkw.WKWDataFormat
-import com.scalableminds.webknossos.datastore.dataformats.zarr.ZarrLayer
-import com.scalableminds.webknossos.datastore.dataformats.zarr3.Zarr3Layer
 import com.scalableminds.webknossos.datastore.helpers.IntervalScheduler
 import com.scalableminds.webknossos.datastore.models.datasource._
 import com.scalableminds.webknossos.datastore.models.datasource.inbox.{InboxDataSource, UnusableDataSource}
